Google_ADK/requirements.txt
google-adk[extensions]
google-adk